requirements.txt
python-dotenv
google-genai
streamlit
yt-dlp
requests